Finished testing part of driver for Cilk Plus; still need timing part

+++ list.hh
@@ -15,7 +15,7 @@ typedef long index_t; //!< Node 'address': -1 (NIL), 0, 1, 2, 3, ...
 #define NIL -1 //!< Index equivalent of a NULL pointer
 
 /** Returns a newly allocated copy of an array */
-index_t* duplicate (int n, const index_t* A);
+index_t* duplicate (size_t n, const index_t* A);
 
 /**
  *  Allocates a pool of 'next' pointers, and initializes into a single
@@ -23,7 +23,7 @@ index_t* duplicate (int n, const index_t* A);
  *  the tail is the element whose next pointer is -1 (i.e., the
  *  element 'k' such that 'next[k] == NIL').
  */
-index_t* createRandomList (int n);
+index_t* createRandomList (size_t n);
 
 /** Frees list array pool. */
 void releaseListBuffer (index_t* Next);
